+++ language.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Text
+from sqlalchemy import Column, Integer, Text, BigInteger
 
 from base import Base
 
@@ -6,5 +6,5 @@ from base import Base
 class Language(Base):
     __tablename__ = "languages"
 
-    tele_id = Column(Integer, primary_key=True)
+    tele_id = Column(BigInteger, primary_key=True)
     language = Column(Text)
