+++ player.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Text, PickleType, ForeignKey
+from sqlalchemy import Column, Integer, Text, PickleType, ForeignKey, BigInteger
 
 from base import Base
 
@@ -6,8 +6,8 @@ from base import Base
 class Player(Base):
     __tablename__ = "players"
 
-    group_tele_id = Column(Integer, ForeignKey("games.group_tele_id"))
-    player_tele_id = Column(Integer, primary_key=True)
+    group_tele_id = Column(BigInteger, ForeignKey("games.group_tele_id"))
+    player_tele_id = Column(BigInteger, primary_key=True)
     player_name = Column(Text)
     player_id = Column(Integer)
     cards = Column(PickleType)
